Quick update
Added a csv file with the data that is being read when added to the file pool in opensesame. Added normalized levenshtein distance, when LD above 0.8 but not correct answer is displayed as orange instead of red. Still need to make some changes so that we can select which experiment we want to run.